refactor(solidjs-tailwind): refactors className files to css modules
