--- solidjs-tailwind/src/components/Header/Header.classNames.js
@@ -1,2 +1,0 @@
-export const header = 'bg-gray-900 flex justify-between items-center py-4 px-8';
-export const navLink = 'text-white font-semibold text-lg';
