+++ solidjs-tailwind/src/components/Header/Header.jsx
@@ -4,32 +4,32 @@ import { useAuth } from '../../auth';
 import { UserDropdown } from '../UserDropdown';
 
 import { GithubLogo } from './GithubLogo';
-import * as styles from './Header.classNames';
+import styles from './Header.module.css';
 
 const Header = () => {
   const { authStore } = useAuth();
 
-    return (
-      <Show when={authStore.user}>
-        <header class={styles.header}>
-          <NavLink href="/">
-            <GithubLogo />
-          </NavLink>
-          <div>
-            {authStore.user ? (
-              <UserDropdown
-                image={authStore.user.avatarUrl}
-                username={authStore.user.login}
-              />
-            ) : (
-              <NavLink href="/api/auth/signin">
-                <span class={styles.navLink}>Sign In</span>
-              </NavLink>
-            )}
-          </div>
-        </header>
-      </Show>
-    );
+  return (
+    <Show when={authStore.user}>
+      <header class={styles.header}>
+        <NavLink href="/">
+          <GithubLogo />
+        </NavLink>
+        <div>
+          {authStore.user ? (
+            <UserDropdown
+              image={authStore.user.avatarUrl}
+              username={authStore.user.login}
+            />
+          ) : (
+            <NavLink href="/api/auth/signin">
+              <span class={styles.navLink}>Sign In</span>
+            </NavLink>
+          )}
+        </div>
+      </header>
+    </Show>
+  );
 };
 
 export default Header;
