+++ solidjs-tailwind/src/components/Header/Header.module.css
@@ -1,0 +1,9 @@
+@tailwind utilities;
+
+.header {
+  @apply bg-gray-900 flex justify-between items-center py-4 px-8;
+}
+
+.navLink {
+  @apply text-white font-semibold text-lg;
+}
