--- solidjs-tailwind/src/components/RepoMeta/RepoMeta.classNames.js
@@ -1,4 +1,0 @@
-export const metadata = 'flex  mt-4 text-xs text-gray-600 space-x-4';
-export const languageColor = 'w-3 h-3 inline-block rounded-full mr-1 relative top-0.5';
-export const socialCount = 'hover:cursor-pointer hover:text-blue-600';
-export const socialIcon = 'w-4 h-4 inline mb-0.5';
