+++ solidjs-tailwind/src/components/RepoMeta/RepoMeta.module.css
@@ -1,0 +1,17 @@
+@tailwind utilities;
+
+.metadata {
+  @apply flex  mt-4 text-xs text-gray-600 space-x-4;
+}
+
+.languageColor {
+  @apply w-3 h-3 inline-block rounded-full mr-1 relative top-0.5;
+}
+
+.socialCount {
+  @apply hover:cursor-pointer hover:text-blue-600;
+}
+
+.socialIcon {
+  @apply w-4 h-4 inline mb-0.5;
+}
