--- solidjs-tailwind/src/components/TabNavigation/tab-navigation.classNames.js
@@ -1,8 +1,0 @@
-export const container = 'border-b border-gray-200';
-export const nav = '-mb-px flex';
-export const tab = 'inline-flex items-center p-4 border-b-2 font-medium text-sm';
-export const tabInactive = 'border-transparent text-gray-600 hover:border-gray-300';
-export const tabActive = 'border-yellow-500 text-gray-900 font-semibold';
-export const icon = '-ml-0.5 mr-2 h-5 w-5 border-none';
-export const iconInactive = 'text-gray-500';
-export const iconActive = 'text-gray-700';
