--- solidjs-tailwind/src/components/TopRepos/TopRepos.classNames.js
@@ -1,6 +1,0 @@
-export const container = 'w-full border rounded-lg relative bg-white';
-export const item = 'p-4 border-b';
-export const headingLink = 'text-xl text-blue-600 font-semibold hover:underline mr-3';
-export const description = 'text-gray-600 text-sm max-w-prose -mb-1 -mt-1';
-export const linkContainer = 'bg-gray-50 p-5 w-full text-center';
-export const allRepoLink = 'font-semibold text-gray-600 hover:text-blue-500';
