+++ solidjs-tailwind/src/components/UserDropdown/UserDropdown.jsx
@@ -2,7 +2,7 @@ import { NavLink } from '@solidjs/router';
 import { createSignal } from 'solid-js';
 import { ChevronDownIcon } from '@heroicons/react/solid';
 import { clickOutside } from '../../utils/onclick-outside';
-import * as styles from './user-dropdown.classNames';
+import styles from './UserDropdown.module.css';
 
 const UserDropdown = (props) => {
   const [expanded, setExpanded] = createSignal(false);
@@ -41,10 +41,7 @@ const UserDropdown = (props) => {
           <ul class="py-1">
             {props.username && (
               <li data-menu-item>
-                <NavLink
-                  href={`/${props.username}`}
-                  class={styles.menuBtn}
-                >
+                <NavLink href={`/${props.username}`} class={styles.menuBtn}>
                   Profile
                 </NavLink>
               </li>
