--- solidjs-tailwind/src/components/UserDropdown/user-dropdown.classNames.js
@@ -1,10 +1,0 @@
-export const dropdown = 'relative top-1 inline-block text-left z-30 relative';
-export const dropdownBtn =
-  'inline-flex items-center text-gray-200 fill-current';
-export const avatarContainer =
-  'w-8 h-8 rounded-full overflow-hidden flex items-center justify-center bg-gray-200';
-export const dropdownMenuBase =
-  'z-40 origin-top-right absolute right-0 mt-2 w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none transition-all duration-200';
-export const dropdownMenuHidden = `${dropdownMenuBase} scale-0 opacity-0`;
-export const menuBtn =
-  'block font-medium px-4 py-2 text-gray-900 hover:text-blue-600';
