--- solidjs-tailwind/src/components/UserProfileCard/user-profile-card.classNames.js
@@ -1,12 +1,0 @@
-export const avatar = 'rounded-full shadow z-30';
-export const name = 'text-2xl text-gray-800 font-bold leading-tight';
-export const username = 'text-xl text-gray-500 font-light';
-export const bio = 'text-gray-800 mt-4 mb-3';
-export const icon = 'w-4 h-4 mb-0.5 mr-1 inline-block';
-export const count = 'font-medium text-gray-900';
-export const socials = 'text-sm text-gray-600 my-4';
-export const fields = 'text-sm text-gray-800 space-y-1';
-export const link = 'hover:text-blue-600 hover:underline';
-export const buttonContainer =
-  'flex flex-col w-4/5 bg-gray-100 mb-3 items-center text-grey-600 rounded-md border-solid border-slate-500 border hover:bg-gray-200';
-export const button = 'border-none text-black py-0.5';
